+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.0
+
+* fixed: cargo test failed
+* Route class now is serializable with serde
+
 # 0.1.2
 
 * add code block type for documents